minecraft - show() fix for Jessie
When wmctrl activates the minecraft window, as of Jessie, it does not
steal mouse focus.  To do that, we either have to hit escape, or click in
the window.  We click ESC twice, to go back to the main menu and return
to the game.

A little hokie, but we don't have another reliable method.